requirements.txt
fastapi
uvicorn
python-multipart
python-docx
reportlab
Pillow
aiofiles
jinja2
